chore: fix circular-imports for suite-desktop-core
