+++ packages/suite-desktop-core/src/modules/external-links.ts
@@ -5,8 +5,7 @@ import { dialog, shell } from 'electron';
 import { HandlerDetails } from 'electron/main';
 
 import * as config from '../config';
-
-import type { ModuleInit } from './index';
+import type { ModuleInit } from './module';
 
 export const SERVICE_NAME = 'external-links';
