+++ packages/suite-desktop-core/src/modules/index.ts
@@ -1,10 +1,8 @@
 import path from 'path';
 
 import { isDevEnv } from '@suite-common/suite-utils';
-import type { DeviceEvent, FirmwareStoreEvent, LocalFirmwares } from '@trezor/connect';
-import { InterceptedEvent } from '@trezor/request-manager';
-import type { HandshakeClient, TorStatus } from '@trezor/suite-desktop-api';
-import { TypedEmitter, isNotUndefined } from '@trezor/utils';
+import type { HandshakeClient } from '@trezor/suite-desktop-api';
+import { isNotUndefined } from '@trezor/utils';
 
 import * as autoStart from './auto-start';
 import * as autoUpdater from './auto-updater';
@@ -17,7 +15,6 @@ import * as customProtocols from './custom-protocols';
 import * as devTools from './dev-tools';
 import * as eventLogging from './event-logging';
 import { MainWindowProxy } from '../libs/main-window-proxy';
-import type { Store } from '../libs/store';
 import * as eventLoggingApp from './event-logging/app';
 import * as eventLoggingContents from './event-logging/contents';
 import * as eventLoggingProcess from './event-logging/process';
@@ -27,6 +24,7 @@ import * as firmware from './firmware';
 import * as httpReceiverModule from './http-receiver';
 import * as menu from './menu';
 import * as metadata from './metadata';
+import { Dependencies, ModuleInit, ModuleInitBackground, ModuleLoad, ModuleQuit } from './module';
 import * as powerMonitor from './power-monitor';
 import * as requestFilter from './request-filter';
 import * as requestInterceptor from './request-interceptor';
@@ -40,6 +38,9 @@ import * as userData from './user-data';
 import * as windowControls from './window-controls';
 
 // General modules (both dev & prod)
+
+export * from './module';
+
 const MODULES: Module[] = [
     // Event Logging
     eventLogging,
@@ -74,46 +75,6 @@ const MODULES: Module[] = [
 ];
 
 const MODULES_BACKGROUND: ModuleBackground[] = [bridge, trezorConnect, httpReceiverModule, tray];
-
-// define events internally sent between modules
-interface MainThreadMessages {
-    'module/request-interceptor': InterceptedEvent;
-    'module/reset-tor-circuits': Extract<InterceptedEvent, { type: 'CIRCUIT_MISBEHAVING' }>;
-    'module/tor-status-update': TorStatus;
-    'module/trezor-connect/device-event': DeviceEvent;
-    'module/bridge/toggle': void;
-    'module/bridge/status': {
-        service: boolean;
-        process: boolean;
-    };
-    'module/trezor-connect/firmware-store': FirmwareStoreEvent;
-    'module/firmware/list': LocalFirmwares;
-    'app/fully-quit': void;
-    'app/show': void;
-}
-
-export const mainThreadEmitter = new TypedEmitter<MainThreadMessages>();
-export type MainThreadEmitter = typeof mainThreadEmitter;
-
-export type Dependencies = {
-    mainWindowProxy: MainWindowProxy;
-    store: Store;
-    interceptor: RequestInterceptor;
-    mainThreadEmitter: MainThreadEmitter;
-};
-
-type ModuleLoad = (payload: HandshakeClient) => any | Promise<any>;
-
-type ModuleQuit = () => void | Promise<void>;
-
-type ModuleInterface = {
-    onLoad: ModuleLoad;
-    onQuit?: ModuleQuit;
-} | void;
-
-export type ModuleInit = (dependencies: Dependencies) => ModuleInterface;
-
-export type ModuleInitBackground = (dependencies: Dependencies) => ModuleInterface;
 
 export type Module = { SERVICE_NAME: string; init: ModuleInit };
 export type ModuleBackground = { SERVICE_NAME: string; initBackground: ModuleInitBackground };
