+++ packages/suite-desktop-core/src/modules/menu.ts
@@ -1,9 +1,8 @@
 import { Menu } from 'electron';
 
+import type { ModuleInit } from './module';
 import { buildMainMenu, inputMenu, selectionMenu } from '../libs/menu';
 import { b2t } from '../libs/utils';
-
-import type { ModuleInit } from './index';
 
 export const SERVICE_NAME = 'menu';
