+++ packages/suite-desktop-core/src/modules/request-filter.ts
@@ -6,8 +6,7 @@ import { captureMessage } from '@sentry/electron/main';
 import { isWhitelistedHost } from '@trezor/utils';
 
 import { allowedDomains, silentlyBlockedDomains } from '../config';
-
-import type { ModuleInit } from './index';
+import type { ModuleInit } from './module';
 
 export const SERVICE_NAME = 'request-filter';
