+++ packages/suite-desktop-core/src/modules/store.ts
@@ -1,6 +1,5 @@
 import { ipcMain } from '../typed-electron';
-
-import type { ModuleInit } from './index';
+import type { ModuleInit } from './module';
 
 export const SERVICE_NAME = 'store';
