+++ packages/suite-desktop-core/src/modules/user-data.ts
@@ -2,8 +2,7 @@ import { validateIpcMessage } from '@trezor/ipc-proxy';
 
 import * as userData from '../libs/user-data';
 import { ipcMain } from '../typed-electron';
-
-import type { ModuleInit } from './index';
+import type { ModuleInit } from './module';
 
 export const SERVICE_NAME = 'user-data';
