+++ packages/suite-desktop-core/src/modules/window-controls.ts
@@ -4,8 +4,7 @@
 import { isWindows } from '@trezor/env-utils';
 
 import { app, ipcMain } from '../typed-electron';
-
-import type { ModuleInit } from './index';
+import type { ModuleInit } from './module';
 
 export const SERVICE_NAME = 'window-control';
