+++ scripts/circular-dependencies/madge-snapshot.txt
@@ -98,36 +98,10 @@ packages/connect/src/events/index.ts > packages/connect/src/events/call.ts > pac
 packages/ipc-proxy/src/proxy-handler.ts > packages/ipc-proxy/src/validateIpcMessage.ts
 packages/product-components/src/components/SelectAssetModal/SelectAssetModal.tsx > packages/product-components/src/components/SelectAssetModal/AssetItem.tsx
 packages/suite-desktop-core/e2e/support/analytics.ts > packages/suite-desktop-core/e2e/support/fixtures.ts
-packages/suite-desktop-core/src/modules/auto-start.ts > packages/suite-desktop-core/src/modules/index.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/auto-updater.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/bluetooth.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/bridge.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/coinjoin.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/crash-recover.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/csp.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/custom-protocols.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/dev-tools.ts
 packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/event-logging/app.ts
 packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/event-logging/contents.ts
 packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/event-logging/index.ts
 packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/event-logging/process.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/external-links.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/file-protocol.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/firmware.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/http-receiver.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/menu.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/metadata.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/power-monitor.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/request-filter.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/request-interceptor.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/shortcuts.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/store.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/system-settings.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/theme.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/tray.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/trezor-connect.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/user-data.ts
-packages/suite-desktop-core/src/modules/index.ts > packages/suite-desktop-core/src/modules/window-controls.ts
 packages/suite/src/actions/suite/metadataActions.ts > packages/suite/src/actions/suite/metadataProviderActions.ts
 packages/suite/src/components/suite/graph/TransactionsGraph/TransactionsGraph.tsx > packages/suite/src/components/suite/graph/TransactionsGraph/GraphTooltipAccount.tsx
 packages/suite/src/components/suite/graph/TransactionsGraph/TransactionsGraph.tsx > packages/suite/src/components/suite/graph/TransactionsGraph/GraphTooltipDashboard.tsx
